requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.6
pymssql==2.2.7
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
